rssmanager/requirements.txt
feedparser
flask
requests
lxml
pillow
